+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- update @cinerino/sdk
+- update @motionpicture/ttts-api-nodejs-client
+
 ### Deprecated
 
 ### Removed